MUSIC IS BACK BAYBEEE
After thinking long and hard about either composing original music for the game, or searching for a guaranteed copyright-free song to use, I have decided to make an arrangement of Beethoven's 'Pathetique Sonata'. Though I will probably want to add more verses of this song later so it doesn't get boring to listen to so quickly. (Alternatively, perhaps more music is in order?)