optimizations?
fixed bug that causes gun to jam by decrementing the cooldown meter in _physics_process.

(probably) optimized the enemy spawning function by declaring the random number int on start rather than re-declaring it in a for loop. That probably caused the garbage collector to freak out out hence why the first spawn wave always hangs the game for a second.